refector: room 관리 관련 event의 logic을 service로 이동
